Collect artifact size metrics after caching the content
The caching machinery stores all headers from the response coming from
content-handler. With this commit, we are introducing an proprietary header
that contains the size of the served artifact or artifact available after the
redirect, X-PULP-ARTIFACT-SIZE.

closes #5817

(cherry picked from commit e709de5c91629e4a77daf0160f8c8f4fd4cef66b)
