+++ CHANGES/5817.bugfix
@@ -1,0 +1,1 @@
+Started collecting artifact size metrics even after caching the served content.
